Increased size of ofs totals to stop overflow

+++ Drive/drive_main/drive_ofs.hpp
@@ -74,11 +74,11 @@ class drive_ofs {
     byte frame[ADNS3080_PIXELS_X * ADNS3080_PIXELS_Y];
 
 public:
-    int total_x = 0; // mm
-    int total_y = 0; // mm
+    long long int total_x = 0; // mm
+    long long int total_y = 0; // mm
 
-    int total_x1 = 0; // pixels
-    int total_y1 = 0; // pixels
+    long long int total_x1 = 0; // pixels
+    long long int total_y1 = 0; // pixels
 
     drive_ofs():total_x(0), total_y(0), total_x1(0), total_y1(0), x(0), y(0), a(0), b(0), distance_x(0), distance_y(0), tdistance(0), movementflag(0){}
     int convTwosComp(int b);
